euler_34

+++ src/aux_funcs.rs
@@ -1,10 +1,10 @@
-pub fn int_to_digits(n: u64) -> Vec<u8> {
+pub fn int_to_digits(n: u64) -> Vec<u64> {
     let mut digits = Vec::new();
     let mut num = n;
     while num != 0 {
         let q = num/10;
         let r = num%10;
-        digits.insert(0,r as u8);
+        digits.insert(0,r as u64);
         num = q;
     }
     return digits;
